# src/main/webapp/WEB-INF/api/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8085","description":"Generated server url"}],"paths":{"/api-birdspecies/content/all":{"post":{"tags":["content-controller"],"summary":"Create a list of contents","operationId":"createContentList","parameters":[{"name":"key","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Content"}}}},"required":true},"responses":{"200":{"description":"Successfully created a list of text blocks"},"404":{"description":"Object not found"},"500":{"description":"Internal Server Error"}}}},"/api-birdspecies/content/":{"get":{"tags":["content-controller"],"summary":"Get the latest content for a specific partitionKey","operationId":"getContent","parameters":[{"name":"key","in":"query","required":true,"schema":{"type":"string"}},{"name":"id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Found content object","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Content"}}}},"404":{"description":"Object not found"},"500":{"description":"Internal Server Error"}}},"post":{"tags":["content-controller"],"summary":"Delete content object","operationId":"addContent_1","parameters":[{"name":"key","in":"query","required":true,"schema":{"type":"string"}},{"name":"userName","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Content"}}}},"required":true},"responses":{"200":{"description":"Successfully deleted a text block"},"404":{"description":"Object not found"},"500":{"description":"Internal Server Error"}}}},"/api-birdspecies/content/selected/":{"get":{"tags":["content-controller"],"summary":"Get list of content for a specific partitionKey","operationId":"getContentsByUuids","parameters":[{"name":"key","in":"query","required":true,"schema":{"type":"string"}},{"name":"uuids","in":"query","required":true,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"Found content objects","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Content"}}}},"404":{"description":"Object not found"},"500":{"description":"Internal Server Error"}}}},"/api-birdspecies/content/all/":{"get":{"tags":["content-controller"],"summary":"Get the list of the latest content","operationId":"getLatestContent","parameters":[{"name":"key","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Found content","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Content"}}}},"404":{"description":"Object not found"},"500":{"description":"Internal Server Error"}}}}},"components":{"schemas":{"Content":{"required":["commitNumber","contentType","created","createdBy","uuid","versionNumber"],"type":"object","properties":{"versionNumber":{"type":"integer","format":"int32"},"commitNumber":{"type":"integer","format":"int32"},"created":{"type":"integer","format":"int64"},"createdBy":{"type":"string"},"uuid":{"type":"string"},"contentType":{"type":"string","enum":["TEXT","LIST","TABLE","ATTACHMENT_REFERENCE","PICTURE_REFERENCE","REFERENCE","HEADER"]},"headerContent":{"$ref":"#/components/schemas/HeaderContent"},"textContent":{"$ref":"#/components/schemas/TextContent"},"listContent":{"$ref":"#/components/schemas/ListContent"},"tableContent":{"$ref":"#/components/schemas/TableContent"}}},"HeaderContent":{"required":["headerType","text"],"type":"object","properties":{"text":{"type":"string"},"headerType":{"type":"string","enum":["ONE","TWO","THREE","FOUR","FIVE"]}}},"ListContent":{"required":["listType","rows"],"type":"object","properties":{"listType":{"type":"string","enum":["NUMBERED","BULLETED"]},"rows":{"type":"array","items":{"type":"string"}}}},"TableContent":{"required":["tableHeader"],"type":"object","properties":{"tableHeader":{"type":"string"},"columnHeaders":{"type":"array","items":{"type":"string"}},"rows":{"type":"array","items":{"$ref":"#/components/schemas/TableRow"}}}},"TableRow":{"required":["celldata"],"type":"object","properties":{"celldata":{"type":"array","items":{"type":"string"}}}},"TextContent":{"required":["text"],"type":"object","properties":{"text":{"type":"string"}}}}}}